add max_num_points argument - thin before plotting
This is my first time extending ggplot2, so apologies if I’ve done anything nonstandard

This commit deals with the problem of large file size when a ROC plot based on a large dataset is stored as a vector graphic.

The solution implemented here is to downsample the data at the end of StatROC (not the beginning!).  In this way, the user gets the benefit of having a lot of data in terms of precision of ROC, but file sizes will stay manageable.  There’s no benefit in plotting more that 1e3 (the new default max_num_points) along a ROC curve.  No one can see that many points.

Signed-off-by: Robert Corty <rcorty@gmail.com>